Rewrote directory structure
