+++ deps.ts
@@ -1,11 +1,17 @@
-export { dirname, join, parse, resolve } from "https://deno.land/std@0.147.0/path/mod.ts";
-export { ensureDir, ensureSymlink, walk } from "https://deno.land/std@0.147.0/fs/mod.ts";
+export {
+  dirname,
+  join,
+  parse,
+  resolve,
+} from "https://deno.land/std@0.147.0/path/mod.ts";
+export {
+  ensureDir,
+  ensureSymlink,
+  walk,
+} from "https://deno.land/std@0.147.0/fs/mod.ts";
 
 export { parseBuffer } from "https://deno.land/x/bplist_parser@0.2.1/mod.ts";
 export { cryptoRandomString as crStr } from "https://deno.land/x/crypto_random_string@1.1.0/mod.ts";
-
-import arch from "https://deno.land/x/arch@v1.0.0/mod.ts";
-export { arch };
 
 import dir from "https://deno.land/x/dir@1.4.0/mod.ts";
 export { dir };
