+++ mod.ts
@@ -1,4 +1,3 @@
-export * from "./lib/alfredo/typed.d.ts";
 export { withPermissions } from "./lib/utils/permissions.ts";
-export { Alfredo, alfredo } from "./lib/alfredo/mod.ts";
+export * as alfredo from "./lib/alfredo/mod.ts";
 export const VERSION = "0.3.0";
